Created a spatial pooler test bench starting from Ian's code.
The goal is to make the top level as concise and easy to understand
as possible.  Right now the classifier is a complete hack, but it
does let me know when the SP fails to distinguish between unique
input patterns.

The next step, I think, is to encode the ground truth tag as part
of the input vector during the training phase and then figure out
how to recover it during the testing phase.